creating intermediary modal to sit between login and actual access to application --> will make sure that modal does not redisplay while navigating around hub routes
