+++ frontend/app/(login)/login/page.tsx
@@ -6,6 +6,30 @@ import {animated, useTransition} from "@react-spring/web";
 import styles from "@/styles/styles.module.css";
 import Sidebar from "@/components/sidebar";
 import Box from "@mui/joy/Box";
+import {
+  useAttributeLoadDispatch,
+  useCensusLoadDispatch,
+  usePersonnelLoadDispatch,
+  usePlotsLoadDispatch,
+  useQuadratsLoadDispatch,
+  useSpeciesLoadDispatch,
+  useSubSpeciesLoadDispatch
+} from "@/app/contexts/fixeddatacontext";
+import {
+  Button,
+  DialogActions,
+  DialogContent,
+  DialogTitle,
+  LinearProgress,
+  Modal,
+  ModalDialog,
+  Stack,
+  Typography
+} from "@mui/joy";
+import WarningRoundedIcon from "@mui/icons-material/WarningRounded";
+import Divider from "@mui/joy/Divider";
+import {useFirstLoadContext, useFirstLoadDispatch} from "@/app/contexts/plotcontext";
+import EntryModal from "@/components/client/entrymodal";
 
 const slides = [
   'background-1.jpg',
@@ -29,9 +53,18 @@ export default function Page() {
     },
     exitBeforeEnter: true,
   })
-  useEffect(() => void setInterval(() => setIndex(state => (state + 1) % slides.length), 5000), [])
+  useEffect(
+    () => {
+      void setInterval(() => setIndex(state => (state + 1) % slides.length), 5000)
+    }, []);
   const {status} = useSession();
-  if (status == "authenticated") redirect('/dashboard');
+  if (status == "authenticated") {
+    return (
+      <>
+        <EntryModal />
+      </>
+    );
+  }
   else
     return (
       <>
