+++ frontend/components/client/endpoint.tsx
@@ -32,65 +32,6 @@ import {useFirstLoadContext, useFirstLoadDispatch} from "@/app/contexts/plotcont
 import WarningRoundedIcon from "@mui/icons-material/WarningRounded";
 
 export default function Endpoint({children,}: { children: React.ReactNode }) {
-  const [loading, setLoading] = useState(0);
-  const [loadingMsg, setLoadingMsg] = useState('');
-  const attributeLoadDispatch = useAttributeLoadDispatch();
-  const censusLoadDispatch = useCensusLoadDispatch();
-  const personnelLoadDispatch = usePersonnelLoadDispatch();
-  const quadratsLoadDispatch = useQuadratsLoadDispatch();
-  const speciesLoadDispatch = useSpeciesLoadDispatch();
-  const plotsLoadDispatch = usePlotsLoadDispatch();
-  useEffect(() => {
-    const fetchData = async () => {
-      setLoading(0);
-      setLoadingMsg('Retrieving Attributes...');
-      let response = await fetch(`/api/fixeddata/attributes`, {method: 'GET'});
-      setLoading(9);
-      if (attributeLoadDispatch) {
-        attributeLoadDispatch({attributeLoad: await response.json()});
-      }
-      setLoading(18);
-      setLoadingMsg('Retrieving Census...');
-      response = await fetch(`/api/fixeddata/census`, {method: 'GET'});
-      setLoading(27);
-      if (censusLoadDispatch) {
-        censusLoadDispatch({censusLoad: await response.json()});
-      }
-      setLoading(36);
-      setLoadingMsg('Retrieving Personnel...');
-      await new Promise(f => setTimeout(f, 1000));
-      // response = await fetch(`/api/fixeddata/personnel`, {method: 'GET'});
-      setLoading(45);
-      // if(personnelLoadDispatch){
-      //   personnelLoadDispatch({personnelLoad: await response.json()});
-      // }
-      setLoading(54);
-      setLoadingMsg('Retrieving Quadrats...');
-      await new Promise(f => setTimeout(f, 1000));
-      // response = await fetch(`/api/fixeddata/quadrats`, {method: 'GET'});
-      setLoading(63);
-      // if (quadratsLoadDispatch) {
-      //   quadratsLoadDispatch({quadratsLoad: await response.json()});
-      // }
-      setLoading(72);
-      setLoadingMsg('Retrieving Species...');
-      await new Promise(f => setTimeout(f, 1000));
-      // response = await fetch(`/api/fixeddata/species`, {method: 'GET'});
-      setLoading(81);
-      // if (speciesLoadDispatch) {
-      //   speciesLoadDispatch({speciesLoad: await response.json()});
-      // }
-      setLoading(90)
-      setLoadingMsg('Retrieving Plots...')
-      response = await fetch(`/api/fixeddata/plots`, {method: 'GET'});
-      setLoading(99);
-      if (plotsLoadDispatch) {
-        plotsLoadDispatch({plotsLoad: await response.json()});
-      }
-      setLoading(100);
-    }
-    fetchData().catch(console.error);
-  }, []);
   useSession({
     required: true,
     onUnauthenticated() {
@@ -133,8 +74,6 @@ export default function Endpoint({children,}: { children: React.ReactNode }) {
   }
   
   let pathname = usePathname();
-  const firstLoad = useFirstLoadContext();
-  const firstLoadDispatch = useFirstLoadDispatch();
   return (
     <>
       <Sidebar/>
@@ -173,41 +112,7 @@ export default function Endpoint({children,}: { children: React.ReactNode }) {
             flexDirection: 'column',
             paddingLeft: 2
           }}>
-          {firstLoad ? <Modal open={firstLoad}
-                              sx={{display: 'flex', flex: 1}}
-                              onClose={(_event: React.MouseEvent<HTMLButtonElement>, reason: string) => {
-                                if (reason !== 'backdropClick' && reason !== 'escapeKeyDown') {
-                                  firstLoadDispatch ? firstLoadDispatch({firstLoad: false}) : null
-                                }
-                              }}>
-            <ModalDialog variant="outlined" role="alertdialog">
-              <DialogTitle>
-                <WarningRoundedIcon/>
-                <Typography level={"title-lg"}>Welcome to the Application!</Typography>
-              </DialogTitle>
-              <Divider/>
-              <DialogContent>
-                <Stack direction={"column"} sx={{display: 'flex', flex: 1}}>
-                  {loading != 100 ?
-                    <>
-                      <LinearProgress sx={{display: 'flex', flex: 1}} determinate size={"lg"} value={loading}/>
-                      <Typography level="body-sm" color="neutral"><b>{loadingMsg}</b></Typography>
-                    </> :
-                    <>
-                      <Typography level={"body-sm"} >Select <b>Core Measurements Hub</b> to view existing core measurement data for a given plot, census, and quadrat</Typography>
-                      <Typography level={"body-sm"}>Select <b>CSV & ArcGIS File Upload Hub</b> to upload core measurements in either CSV format or in collected ArcGIS format</Typography>
-                      <Typography level={"body-sm"}>Select <b>Measurement Properties Hub</b> to view and edit measurement properties used in data collection</Typography>
-                    </>}
-                </Stack>
-              </DialogContent>
-              <DialogActions>
-                <Button variant="plain" color="neutral" disabled={loading != 100}
-                        onClick={() => firstLoadDispatch ? firstLoadDispatch({firstLoad: false}) : null}>
-                  Continue
-                </Button>
-              </DialogActions>
-            </ModalDialog>
-          </Modal> : children}
+          {children}
         </Box>
         <Divider orientation={"horizontal"}/>
         <Box mt={3}
